refactor package and worked pull request
