+++ notes/src/main/java/ru/enedinae/notes/app/Application.java
@@ -1,12 +1,12 @@
 package ru.enedinae.notes.app;
 
-import ru.enedinae.notes.logic.CommandLineUiImpl;
+import ru.enedinae.notes.ui.UserInterface;
+import ru.enedinae.notes.ui.impl.CommandLineUiImpl;
 
 public class Application {
-    //private UserInterface ui;
+    private UserInterface ui;
     public static void main(String[] args)  {
+
         new CommandLineUiImpl().start();
     }
 }
-
-
