+++ notes/src/main/java/ru/enedinae/notes/enumeration/NoteStatus.java
@@ -1,4 +1,4 @@
-package ru.enedinae.notes;
+package ru.enedinae.notes.enumeration;
 
 public enum NoteStatus {
     NEW, CLOSED
