+++ notes/src/main/java/ru/enedinae/notes/model/NotesIdsGenerator.java
@@ -1,4 +1,4 @@
-package ru.enedinae.notes.logic;
+package ru.enedinae.notes.model;
 
 public class NotesIdsGenerator {
     private int id = 0;
