+++ notes/src/main/java/ru/enedinae/notes/service/impl/InMemoryNotesServiceImpl.java
@@ -1,10 +1,11 @@
-package ru.enedinae.notes.logic;
+package ru.enedinae.notes.service.impl;
 
-import ru.enedinae.notes.NotesService;
-import ru.enedinae.notes.models.Note;
+import ru.enedinae.notes.model.Note;
+import ru.enedinae.notes.service.NotesService;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 public class InMemoryNotesServiceImpl implements NotesService {
@@ -17,27 +18,30 @@ public class InMemoryNotesServiceImpl implements NotesService {
     }
 
     public List<Note> getAllNotes() {
+
         return notes;
     }
 
     public Optional<Note> getNoteById(Integer id) {
-        for(Note note : notes) {
+        return notes.stream().filter(note -> Objects.equals(note.getId(), id)).findFirst();
+        /*for(Note note : notes) {
             Optional<Integer> op = Optional.ofNullable(note.getId());
             if(op.isPresent() && note.getId().equals(id)) {
                 return Optional.of(note);
             }
         }
-        return Optional.empty();
+        return Optional.empty();*/
     }
 
     public Optional<Note> getNoteByName(String name) {
-        for(Note note : notes) {
+        return notes.stream().filter(note -> Objects.equals(note.getName(), name)).findFirst();
+        /*for(Note note : notes) {
             Optional<String> op = Optional.ofNullable(note.getName());
             if(op.isPresent() && note.getName().equals(name)) {
                 return Optional.of(note);
             }
         }
-        return Optional.empty();
+        return Optional.empty();*/
     }
 
     public boolean deleteNoteById(Integer id) {
@@ -49,12 +53,6 @@ public class InMemoryNotesServiceImpl implements NotesService {
     }
 
     public boolean updateNote(Note updateNote) {
-        for (Note note : notes) {
-            if(note.getId().equals(updateNote.getId())) {
-                note = updateNote;
-                return true;
-            }
-        }
-        return false;
+        return true;
     }
 }
