+++ notes/src/main/java/ru/enedinae/notes/ui/UserInterface.java
@@ -1,4 +1,4 @@
-package ru.enedinae.notes;
+package ru.enedinae.notes.ui;
 
 public interface UserInterface {
     void start();
